% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lanczos.R
\name{lanczos_spec}
\alias{lanczos_spec}
\title{Compute Lanczos filter amplitude and phase (not unwound) spectra}
\usage{
lanczos_spec(w, nPts)
}
\arguments{
\item{w}{filter coefficients returned by \code{\link{lanczos}}}

\item{n}{zero-padded length of filter to transform}
}
\value{
data.frame containing amplitude and phase spectrum of filter for non-negative frequencies (DC to Nyquist).
 \item{"freq"}{Fourier frequencies (0 to Nyquist)}
 \item{"amp"}{Fourier amplitude spectrum of filter}
 \item{"phase"}{Fourier phase spectrum of filter}
}
\description{
Compute Lanczos filter amplitude spectrum
}
\examples{
f <- lanczos_bp( 7, 0.1, 0.4 )
s <- lanczos_spec( f, 512 )
plot( s$freq, s$amp, type = "l", xlab = "Frequency", ylab = "Amplitude" )

}
